bc-pulscen-main/requirements.txt
fastapi
uvicorn[standard]
httpx
aiohttp
pydantic
playwright
openai
python-dotenv
pyautogui
Pillow
fpdf
pdfkit
serpapi
pytesseract
requests
